requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
pydantic>=2.0.0
PyYAML>=6.0.0
httpx>=0.27.0
pytest>=8.0.0
Jinja2>=3.1.0
